Refactored the code base  * Changed from table to divs  * Implemented all CRUD operations
