+++ src/index.html
@@ -7,90 +7,50 @@
   <link rel="stylesheet" href="style.css">
   <link rel="stylesheet" href="./styles/div.css">
   <script src="./index.js" defer type="module"></script>
-  <script src="./modules/pop.js" defer type="module"></script>
   <title>My Timetable</title>
 </head>
 <body>
-  <!-- <div class="time-table">
-    <h1 class="heading">Let's account for out time</h1>
-    <table>
-      <caption>The value of time is beyond computation</caption>
-      <thead>
-        <tr>
-          <th scope="col">Sunday</th>
-          <th scope="col">Monday</th>
-          <th scope="col">Tuesday</th>
-          <th scope="col">Wednesday</th>
-          <th scope="col">Thursday</th>
-          <th scope="col">Friday</td>
-          <th scope="col">Saturday</th>
-        </tr>
-      </thead>
-      <tbody>
-        <tr>
-          <td class="sunday">
-            <span class="plus">+</span>
-          </td>
-          <td class="monday">
-            <span class="plus">+</span>
-          </td>
-          <td class="tuesday">
-            <span class="plus">+</span>
-          </td>
-          <td class="wednesday">
-            <span class="plus">+</span>
-          </td>
-          <td class="thursday">
-            <span class="plus">+</span>
-          </td>
-          <td class="friday">
-            <span class="plus">+</span>
-          </td>
-          <td class="saturday">
-            <span class="plus">+</span>
-          </td>
-        </tr>
-      </tbody>
-    </table>
-  </div>
-  <dialog id="favDialog">
-    <div id="form-containerr"></div>
-  </dialog> -->
-
   <dialog id="popUp">
     <div id="form-container"></div>
   </dialog>
-
+  <caption>The Value of Time is beyond computation</caption>
   <div class="container">
+    
     <div class="day-of-the-week">
-      <div class="sunday">Sunday</div>
+        Sunday
+      <div class="sunday"></div>
       <span class="new-task">Add task</span>
     </div>
     <div class="day-of-the-week">
-      <div class="monday">Monday</div>
+      Monday
+      <div class="monday"></div>
       <span class="new-task">Add task</span>
     </div>
     <div class="day-of-the-week">
-      <div class="tuesday">Tuesday</div>
+      Tuesday
+      <div class="tuesday"></div>
       <span class="new-task">Add task</span>
     </div>
     <div class="day-of-the-week">
-      <div class="wednesday">Wednesday</div>
+      Wednesday
+      <div class="wednesday"></div>
       <span class="new-task">Add task</span>
     </div>
     <div class="day-of-the-week">
-      <div class="thursday">Thursday</div>
+      Thursday
+      <div class="thursday"></div>
       <span class="new-task">Add task</span>
     </div>
     <div class="day-of-the-week">
-      <div class="friday">Friday</div>
+      Friday
+      <div class="friday"></div>
       <span class="new-task">Add task</span>
     </div>
     <div class="day-of-the-week">
-      <div class="saturday">Saturday</div>
+      Saturday
+      <div class="saturday"></div>
       <span class="new-task">Add task</span>
     </div>
   </div>
-
 </body>
 </html>
