+++ src/index.js
@@ -1,15 +1,19 @@
-import { handleEdit, handleEditButtons } from './edit.js';
-import displayStoredData from './display.js';
-import { handleDeleteButtons } from './delete.js';
+import { handleEdit, handleEditButtons } from './modules/edit.js';
+import displayStoredData from './modules/display.js';
+import { handleDeleteButtons } from './modules/delete.js';
+import handleSubmitForm from './modules/handleFormSubmit.js';
 
-export const dialog = document.getElementById('favDialog');
+const storedDataString = localStorage.getItem('userData');
+const storedData = storedDataString ? JSON.parse(storedDataString) : {};
+
+export const dialog = document.getElementById('popUp');
 
 export const openCheck = (dialog, className) => {
   if (dialog.open) {
     document.getElementById('form-container').innerHTML = `
       <div>
         <h1>Plan your ${className}</h1>
-        <form method="" id="form-dataa">
+        <form method="" id="form-data">
           <label name="task">Task</label>
           <input type="text" placeholder="Enter task..." id="task">
           <label name="date">Date:</label>
@@ -34,57 +38,29 @@ export const openCheck = (dialog, className) => {
     const handleSubmit = document.getElementById('form-data');
     handleSubmit.addEventListener('submit', (e) => {
       e.preventDefault();
-      const taskInput = document.getElementById('task').value;
-      const dateInput = document.getElementById('date').value;
-      const fromTimeInput = document.getElementById('from-time').value;
-      const toTimeInput = document.getElementById('to-time').value;
-      const tdElement = document.querySelector(`td.${className}`);
-
-      const storedDataString = localStorage.getItem('userData');
-      const storedData = storedDataString ? JSON.parse(storedDataString) : {};
-      const key = `${className}-${Date.now()}`;
-
-      storedData[key] = {
-        task: taskInput,
-        date: dateInput,
-        fromTime: fromTimeInput,
-        toTime: toTimeInput,
-      };
-
-      localStorage.setItem('userData', JSON.stringify(storedData));
-      displayStoredData(tdElement, storedData[key]);
-      // eslint-disable-next-line no-use-before-define
-      handleAdd(tdElement, { [key]: storedData[key] });
-      dialog.close();
+      handleSubmitForm(className, dialog);
     });
   }
 };
 
-const storedDataString = localStorage.getItem('userData');
-const storedData = storedDataString ? JSON.parse(storedDataString) : {};
+// Add elements
+document.addEventListener('click', (event) => {
+  if (event.target.classList.contains('new-task')) {
+    dialog.showModal();
+    openCheck(dialog, event.target.previousElementSibling.className);
+  }
+});
 
-const tableCells = document.querySelectorAll('td');
-displayStoredData(tableCells, storedData);
-
-handleDeleteButtons();
-
+// Edit elements
 document.addEventListener('click', (event) => {
   if (event.target.classList.contains('edit')) {
     const editButton = event.target;
     handleEdit(editButton, dialog, openCheck);
-    handleEditButtons();
+    handleEditButtons(dialog, openCheck);
   }
 });
 
-document.addEventListener('click', (event) => {
-  if (event.target.classList.contains('plus')) {
-    dialog.showModal();
-    openCheck(dialog, event.target.parentNode.className);
-  }
-});
+const timeTableTasks = document.querySelectorAll('.sunday, .monday, .tuesday, .wednesday, .thursday, .friday, .saturday');
+displayStoredData(timeTableTasks, storedData);
 
-export const handleAdd = (tdElement, data) => {
-  displayStoredData(tdElement, data);
-  handleDeleteButtons();
-  handleEditButtons(dialog, openCheck);
-};
+handleDeleteButtons();
